expend folder structure start to create the first step of the sloby-editor
