+++ src/sloby-editor-framework/creators/injector/index.ts
@@ -1,0 +1,3 @@
+export class Injector {
+    constructor() {}
+}
