+++ src/sloby-editor-framework/implementations/SlobyEditor/Editor.tsx
@@ -1,0 +1,16 @@
+import React from "react"
+import { AdminPageContainer } from "../../../Editor/utils/styles/AdminPage"
+import { SlobyEditorContainer, SlobyConstantTools } from "../../../Editor/utils/styles/Editor"
+import SlobyPreviewSite from "../SlobyPreviewSite/SlobyPreviewSite"
+import SlobyTools from "../SlobyTools/SlobyTools"
+
+function SlobyEditor() {
+  return <SlobyEditorContainer>
+      <SlobyConstantTools>
+        <SlobyTools />
+        <SlobyPreviewSite />
+      </SlobyConstantTools>
+  </SlobyEditorContainer>
+}
+
+export default SlobyEditor
