+++ src/sloby-editor-framework/tools/sloby_tools.ts
@@ -1,0 +1,3 @@
+export class SlobyTools {
+  constructor() {}
+}
