+++ src/sloby-editor-framework/utils/types/index.ts
@@ -1,0 +1,3 @@
+export interface SlobyObject {
+  type?: string,
+}
